+++ observable_jupyter/__init__.py
@@ -2,6 +2,7 @@
 
 __author__ = """Thomas Ballinger"""
 __email__ = "me@ballingt.com"
-__version__ = "0.1.14"
+__version__ = "0.1.15"
 
 from .jupyter_embed import embed
+from .jupyter_magics import embed_observable